+++ src/cli.ts
@@ -83,9 +83,6 @@ else if (args.list) {
     else if (Number.isNaN(ver)) {
         console.error(`'${ver}' is not a valid pack format`)
     }
-    else if (!/^\d+$/.test(ver)) {
-        console.error(`'${ver}' is a version number, not a pack format`)
-    }
     else {
         const type = args.data ? 'data' : 'resource'
         const vers = getVersions(+ver, type)